webhook-server/requirements.txt
Flask==3.1.1
gunicorn==23.0.0
python-dotenv==1.1.0
requests-oauthlib==2.0.0
